This code version fixed some issues, but it has a bad fix for an odd observed issue.
The previous issue: Code unresponsive to holding down the right button (although button handler is visibly called and it was hitting my unpause case and setting the tests needed flag to being true.

The half fix: Adding in if not reading on extension microswitch extending flag = true.
This has some issues though because it doesn't activate the three second timer and
once you actually hit the micro it doesn't retract or go into any other state, but it reads the pressure along the way

I think I can fix this by setting the iterator for the ~25 ms "ISR" to be on the pretest logic state (or 1 before depending on if it'll run the code to update the isr counter first.
We could also make a flag for pretest logic itself here and set that to be true.